Update to editor
