+++ MainWindow.xaml
@@ -3,36 +3,79 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:avalonEdit="http://icsharpcode.net/sharpdevelop/avalonedit"
         xmlns:local="clr-namespace:Voice_Coding"
         mc:Ignorable="d"
         
-        Title="MainWindow" Height="128.4" Width="200" ResizeMode="NoResize" WindowStyle="None" Background="#FF9A9A9A" WindowStartupLocation="Manual" Left="0" Top="0" VerticalAlignment="Bottom" HorizontalAlignment="Right" Visibility="Hidden" ShowInTaskbar="False" Icon="VoiceCoding_72.png">
-    <Window.Resources>
-        <Style x:Key="NonFocus" TargetType="ComboBox">
-            <Setter Property="FocusVisualStyle" Value="{x:Null}"/>
-        </Style>
+        Title="Voice Coding" Height="450" Width="800" Icon="Resource/Icon/Voice Coding/32.png" FontFamily="PT Sans">
 
-        <Style x:Key="NonFocusItem" TargetType="ComboBoxItem">
-            <Setter Property="FocusVisualStyle" Value="{x:Null}"/>
-        </Style>
-    </Window.Resources>
-    <Grid>
-        <Rectangle Fill="#FF9A9A9A" HorizontalAlignment="Left" Height="128" VerticalAlignment="Top" Width="200" Stroke="#FFEDEDED" Margin="-1,0,0,0" Grid.RowSpan="2"/>
-
-        <ComboBox x:Name="LanguageSelect" HorizontalAlignment="Left" Margin="11,33,0,0" VerticalAlignment="Top" Width="176" Grid.RowSpan="2">
-            <ComboBoxItem Content="CPP" IsSelected="True"></ComboBoxItem>
-            <ComboBoxItem Content="C" IsSelected="False"></ComboBoxItem>
-            <ComboBoxItem Content="C#" IsSelected="False"></ComboBoxItem>
-        </ComboBox>
-        <Rectangle Fill="#FF313131" HorizontalAlignment="Center" Height="20" VerticalAlignment="Top" Width="198" Margin="1,1,1,0"/>
-        <Button x:Name="CloseBtn" Content="" HorizontalAlignment="Left" Margin="178,1,0,0" VerticalAlignment="Top" Width="21" Height="20" Foreground="#FFD7D7D7" BorderBrush="#FFC75050" Click="HideToTray">
-            <Button.Background>
-                <ImageBrush ImageSource="close.png" Stretch="UniformToFill"/>
-            </Button.Background>
-        </Button>
-        <Label Content="Voice Coding" HorizontalAlignment="Left" VerticalAlignment="Top" Background="{x:Null}" Foreground="#FFD6D6D6" Margin="1,-3,0,0"/>
-        <TextBox x:Name="IncluedPath" HorizontalAlignment="Left" Height="22" Padding="2,2,0,0" Margin="11,62,0,0" Text="Path" VerticalAlignment="Top" Width="176" MaxLines="1"/>
-        <Button x:Name="ReloadBtn" Content="Reload grammar" HorizontalAlignment="Left" Height="22" Margin="48,93,0,0" VerticalAlignment="Top" Width="97"/>
-
-    </Grid>
+    <DockPanel>
+        <ToolBar DockPanel.Dock="Top">
+            <ToolBar.Resources>
+                <Style TargetType="{x:Type Image}">
+                    <Style.Triggers>
+                        <DataTrigger Binding="{Binding RelativeSource={RelativeSource AncestorType={x:Type ButtonBase}, AncestorLevel=1}, Path=IsEnabled}" Value="False">
+                            <Setter Property="Opacity" Value="0.30" />
+                        </DataTrigger>
+                    </Style.Triggers>
+                </Style>
+            </ToolBar.Resources>
+            <Button Click="OpenFileClick">
+                <Image Source="Resource/Icon/Open.png" Height="16"/>
+            </Button>
+            <Button Click="SaveFileClick">
+                <Image Source="Resource/Icon/Save.png" Height="16"/>
+            </Button>
+            <Separator/>
+            <Button Command="Cut">
+                <Image Source="Resource/Icon/Cut.png" Height="16"/>
+            </Button>
+            <Button Command="Copy">
+                <Image Source="Resource/Icon/Copy.png" Height="16"/>
+            </Button>
+            <Button Command="Paste">
+                <Image Source="Resource/Icon/Past.png" Height="16"/>
+            </Button>
+            <Button Command="Delete">
+                <Image Source="Resource/Icon/Delet.png" Height="16"/>
+            </Button>
+            <Separator/>
+            <Button Command="Undo">
+                <Image Source="Resource/Icon/Undo.png" Height="16"/>
+            </Button>
+            <Button Command="Redo">
+                <Image Source="Resource/Icon/Redo.png" Height="16"/>
+            </Button>
+            <Separator/>
+            <CheckBox IsChecked="{Binding ElementName=textEditor,Path=WordWrap}">
+                <Image Source="Resource/Icon/New/Wrap.png" Height="16"/>
+            </CheckBox>
+            <CheckBox IsChecked="{Binding ElementName=textEditor,Path=ShowLineNumbers}">
+                <TextBlock Width="16" TextAlignment="Center">#</TextBlock>
+            </CheckBox>
+            <CheckBox IsChecked="{Binding ElementName=textEditor,Path=Options.ShowEndOfLine}">
+                <TextBlock Width="16" TextAlignment="Center">¶</TextBlock>
+            </CheckBox>
+        </ToolBar>
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="*"/>
+                <RowDefinition Height="27"/>
+            </Grid.RowDefinitions>
+            <Grid Grid.Row="0">
+                <avalonEdit:TextEditor
+				    Name="textEditor"
+				    FontFamily="Consolas"
+				    FontSize="10pt"
+				    SyntaxHighlighting="C++"
+                    ShowLineNumbers="True">
+                    <!--Welcome to AvalonEdit!-->
+                </avalonEdit:TextEditor>
+            </Grid>
+            <Grid Grid.Row="1" Background="#FF373737">
+                <Button x:Name="ToggelButton" Content=""  HorizontalAlignment="Left"  VerticalAlignment="Center" Width="27" UseLayoutRounding="False" Height="27.067" BorderBrush="#FF373737" BorderThickness="3" Click="OnToggleRecognition" Background="#FF40C075"/>
+                <Label x:Name="status" HorizontalAlignment="Left" Margin="27,0,0,-2.8" VerticalAlignment="Top" Foreground="#FFDFDFDF" Height="30" Width="743" FontSize="14" Content="Commands" Background="{x:Null}" FontFamily="Poppins"/>
+            </Grid>
+        </Grid>
+    </DockPanel>
 </Window>
