--- StatusBar.xaml
@@ -1,28 +1,0 @@
-﻿<Window x:Class="Voice_Coding.StatusBar"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        mc:Ignorable="d"
-        Title="StatusBar" Height="25" Width="800" ResizeMode="NoResize" ShowInTaskbar="False" Topmost="True" Cursor="Hand" WindowStyle="None" Foreground="{x:Null}" AllowsTransparency="True" Icon="VoiceCoding_72.png">
-    <Window.Background>
-        <LinearGradientBrush EndPoint="1,0" StartPoint="0,0">
-            <GradientStop Color="#FF373737" Offset="0.801"/>
-            <GradientStop Color="#00320D0D" Offset="1"/>
-        </LinearGradientBrush>
-    </Window.Background>
-    <Window.Resources>
-        <Style TargetType="{x:Type Button}">
-            <Setter Property="FocusVisualStyle" Value="{x:Null}"/>
-        </Style>
-    </Window.Resources>
-    <Grid>
-        <Button x:Name="toggleBtn" Content=""  HorizontalAlignment="Left" Margin="28.731,-0.017,0,0.017" VerticalAlignment="Center" Width="24.6" UseLayoutRounding="False" Height="24.8" BorderBrush="#FF373737" BorderThickness="3" Click="ToggleBtn_Click" Background="#FF40C075" Grid.Column="1"/>
-        <Label x:Name="status" HorizontalAlignment="Left" Margin="53.267,-3,0,-1.2" VerticalAlignment="Top" Foreground="#FFDFDFDF" Height="29" Width="743" FontSize="14" Grid.Column="1" Content="Commands"/>
-        <Button x:Name="ExitStatusBar" HorizontalAlignment="Left" VerticalAlignment="Top" Width="32" Height="27" Foreground="#FFFFEBEB" BorderBrush="{x:Null}" Click="ExitStatusBar_Click" Margin="-1,-1,0,-1.2" FontFamily="Franklin Gothic Demi Cond" Grid.ColumnSpan="2">
-            <Button.Background>
-                <ImageBrush ImageSource="close.png" Stretch="UniformToFill"/>
-            </Button.Background>
-        </Button>
-    </Grid>
-</Window>
