+++ Voice Coding.csproj
@@ -5,7 +5,7 @@
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
     <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
     <ProjectGuid>{A045BE51-D98C-4313-A7A3-467568658C69}</ProjectGuid>
-    <OutputType>WinExe</OutputType>
+    <OutputType>Exe</OutputType>
     <RootNamespace>Voice_Coding</RootNamespace>
     <AssemblyName>Voice Coding</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
@@ -41,9 +41,13 @@
     <StartupObject />
   </PropertyGroup>
   <PropertyGroup>
-    <ApplicationIcon>VoiceCoding_icon_72.ico</ApplicationIcon>
+    <ApplicationIcon>
+    </ApplicationIcon>
   </PropertyGroup>
   <ItemGroup>
+    <Reference Include="ICSharpCode.AvalonEdit, Version=6.0.1.278, Culture=neutral, PublicKeyToken=9cc39be672370310, processorArchitecture=MSIL">
+      <HintPath>packages\AvalonEdit.6.0.1\lib\net45\ICSharpCode.AvalonEdit.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Data" />
     <Reference Include="System.Drawing" />
@@ -61,26 +65,15 @@
     <Reference Include="WindowsBase" />
     <Reference Include="PresentationCore" />
     <Reference Include="PresentationFramework" />
-    <Reference Include="WindowsInput, Version=1.0.4.0, Culture=neutral, PublicKeyToken=9b287f7dc5073cad, processorArchitecture=MSIL">
-      <HintPath>packages\InputSimulator.1.0.4.0\lib\net20\WindowsInput.dll</HintPath>
-    </Reference>
   </ItemGroup>
   <ItemGroup>
     <ApplicationDefinition Include="App.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="DataResource.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>DataResource.resx</DependentUpon>
-    </Compile>
-    <Compile Include="TrayIcon.cs" />
-    <Compile Include="src\CodeRecognition.cs" />
-    <Compile Include="src\CPPGrammar.cs" />
-    <Compile Include="StatusBar.xaml.cs">
-      <DependentUpon>StatusBar.xaml</DependentUpon>
-    </Compile>
+    <Compile Include="Source\TrayIcon.cs" />
+    <Compile Include="Source\CodeRecognition.cs" />
+    <Compile Include="Source\CPPGrammar.cs" />
     <Page Include="MainWindow.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
@@ -93,10 +86,6 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="StatusBar.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Properties\AssemblyInfo.cs">
@@ -112,15 +101,13 @@
       <DependentUpon>Settings.settings</DependentUpon>
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
-    <EmbeddedResource Include="DataResource.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>DataResource.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
     <EmbeddedResource Include="Properties\Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <None Include="packages.config" />
+    <None Include="packages\AvalonEdit.6.0.1\.signature.p7s" />
+    <None Include="packages\AvalonEdit.6.0.1\AvalonEdit.6.0.1.nupkg" />
     <None Include="Properties\Settings.settings">
       <Generator>SettingsSingleFileGenerator</Generator>
       <LastGenOutput>Settings.Designer.cs</LastGenOutput>
@@ -130,22 +117,43 @@
     <None Include="App.config" />
   </ItemGroup>
   <ItemGroup>
-    <None Include="res\Icon\icon_tray_b&amp;w.ico" />
+    <Resource Include="Resource\Images\close.png" />
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="VoiceCoding_72.png" />
+    <Resource Include="Resource\MainResource.xml" />
   </ItemGroup>
   <ItemGroup>
-    <Resource Include="close.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <Resource Include="res\MainResource.xml" />
-  </ItemGroup>
-  <ItemGroup>
-    <Resource Include="VoiceCoding_icon_128.ico" />
-  </ItemGroup>
-  <ItemGroup>
-    <Resource Include="VoiceCoding_icon_72.ico" />
+    <Content Include="packages\AvalonEdit.6.0.1\lib\net40\ICSharpCode.AvalonEdit.dll" />
+    <Resource Include="Resource\Icon\Wrap.png" />
+    <Resource Include="Resource\Icon\Voice Coding\128.png" />
+    <Resource Include="Resource\Icon\Voice Coding\16.png" />
+    <Resource Include="Resource\Icon\Voice Coding\256.png" />
+    <Resource Include="Resource\Icon\Voice Coding\32.png" />
+    <Resource Include="Resource\Icon\Voice Coding\48.png" />
+    <Resource Include="Resource\Icon\Voice Coding\96.png" />
+    <Resource Include="Resource\Icon\Voice Coding\ico\128.ico" />
+    <Resource Include="Resource\Icon\Voice Coding\ico\16.ico" />
+    <Resource Include="Resource\Icon\Voice Coding\ico\32.ico" />
+    <Resource Include="Resource\Icon\Voice Coding\ico\48.ico" />
+    <Resource Include="Resource\Icon\Voice Coding\ico\96.ico" />
+    <Resource Include="Resource\Icon\Voice Coding\ico\Voice_coding_B&amp;W.ico" />
+    <Resource Include="Resource\Icon\Voice Coding\Voice_coding.ico" />
+    <Resource Include="Resource\Icon\Delet.png" />
+    <Resource Include="Resource\Icon\Past.png" />
+    <Content Include="Resource\Icon\Voice Coding\Voice_coding.svg" />
+    <Content Include="Resource\Icon\Voice Coding\Voice_coding_B&amp;W.svg" />
+    <Content Include="Resource\Icon\Voice Coding\Voice_coding_ink.svg" />
+    <Resource Include="Resource\Icon\Copy.png" />
+    <Resource Include="Resource\Icon\Cut.png" />
+    <Resource Include="Resource\Icon\Open.png" />
+    <Resource Include="Resource\Icon\Redo.png" />
+    <Resource Include="Resource\Icon\Save.png" />
+    <Resource Include="Resource\Icon\Undo.png" />
+    <Resource Include="packages\AvalonEdit.6.0.1\lib\net40\ICSharpCode.AvalonEdit.xml" />
+    <Content Include="packages\AvalonEdit.6.0.1\lib\net45\ICSharpCode.AvalonEdit.dll" />
+    <Resource Include="packages\AvalonEdit.6.0.1\lib\net45\ICSharpCode.AvalonEdit.xml" />
+    <Content Include="packages\AvalonEdit.6.0.1\lib\netcoreapp3.0\ICSharpCode.AvalonEdit.dll" />
+    <Resource Include="packages\AvalonEdit.6.0.1\lib\netcoreapp3.0\ICSharpCode.AvalonEdit.xml" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
