+++ packages.config
@@ -1,5 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="AvalonEdit" version="6.0.1" targetFramework="net45" />
-  <package id="InputSimulator" version="1.0.4.0" targetFramework="net472" />
 </packages>
